git subrepo clone --force git://github.com/perl6/perl6-lwp-simple.git perl6/perl6-lwp-simple
subrepo:
  subdir:   "perl6/perl6-lwp-simple"
  merged:   "01b3edf"
upstream:
  origin:   "git://github.com/perl6/perl6-lwp-simple.git"
  branch:   "master"
  commit:   "01b3edf"
git-subrepo:
  version:  "0.4.0"
  origin:   "git@github.com:ingydotnet/git-subrepo.git"
  commit:   "b645cc9"

+++ perl6/perl6-lwp-simple/t/getstore.t
@@ -5,6 +5,12 @@ use LWP::Simple;
 
 plan 4;
 
+if %*ENV<NO_NETWORK_TESTING> {
+    diag "NO_NETWORK_TESTING was set";
+    skip-rest("NO_NETWORK_TESTING was set");
+    exit;
+}
+
 my $fname = $*SPEC.catdir($*TMPDIR, "./tmp-getstore-$*PID");
 try unlink $fname;
